180619 Comparing MAGPHYS and CIGALE sources in the COSMOS field.
This version of the code is set up to compare the ~70 sources we currently have from Nick Scoville. This version focuses on Mdust, Mstar and SFR. This is before we have done the corrections i.e. using the z from MAGPHYS and the smalle cross-matching. Do we need to look into likelihood functions? Probably not high enough density source but worth checking. It produces log-log plots and calculates dispersion values. There are no error considerations at this stage.