180620 Comparing MAGPHYS and CIGALE sources in COSMOS field
This code compares all the matches for sources in the COSMOS field that have both MAGPHYS and CIGALE data. The sources are cut for those which have a very different redshift (for M and C) and by the requirement that there must be detections in two FIR bands (there is no SNR limit at this stage). We have no consideration of errors at this stage. This code is before we have done the corrected work for using the MAGPHYS redshifts and consideration of the number density of sources in matching. The plots created are log-log plots and the dispersion of sources about the m=0 line are calculated.