Comparing Scoville dust masses against MAGPHYS and CIGALE dust masses
This code compares the dust masses calculated from CIGALE and MAGPHYS against dust masses calculated using the Scoville method (850um). This version of the code is only comparing ~70 sources but we have data on ~700 now to improve this work. We make log-log plots for M*, SFR, Mdust and calculate the dispersion about the residuals of a fitted m=0 line.